Rename std/io/{io,}util.ts to std/_util/io.ts
