+++ std/io/mod.ts
@@ -1,4 +1,3 @@
 export * from "./bufio.ts";
-export * from "./ioutil.ts";
 export * from "./readers.ts";
 export * from "./writers.ts";
