+++ std/io/readers_test.ts
@@ -3,7 +3,7 @@ import { test } from "../testing/mod.ts";
 import { assertEquals } from "../testing/asserts.ts";
 import { MultiReader, StringReader } from "./readers.ts";
 import { StringWriter } from "./writers.ts";
-import { copyN } from "./ioutil.ts";
+import { copyN } from "../_util/io.ts";
 import { decode } from "../strings/mod.ts";
 
 test(async function ioStringReader(): Promise<void> {
