+++ std/mime/multipart.ts
@@ -5,11 +5,10 @@ const { min, max } = Math;
 type Closer = Deno.Closer;
 type Reader = Deno.Reader;
 type Writer = Deno.Writer;
+import { copyN, tempFile } from "../_util/io.ts";
 import { equal, findIndex, findLastIndex, hasPrefix } from "../bytes/mod.ts";
-import { copyN } from "../io/ioutil.ts";
 import { MultiReader } from "../io/readers.ts";
 import { extname } from "../path/mod.ts";
-import { tempFile } from "../io/util.ts";
 import { BufReader, BufWriter, UnexpectedEOFError } from "../io/bufio.ts";
 import { encoder } from "../strings/mod.ts";
 import { assertStrictEq } from "../testing/asserts.ts";
