+++ std/textproto/mod.ts
@@ -3,8 +3,8 @@
 // Use of this source code is governed by a BSD-style
 // license that can be found in the LICENSE file.
 
+import { charCode } from "../_util/io.ts";
 import { BufReader, UnexpectedEOFError } from "../io/bufio.ts";
-import { charCode } from "../io/util.ts";
 
 const asciiDecoder = new TextDecoder();
 function str(buf: Uint8Array): string {
